requirements.txt
streamlit
google-generativeai
moviepy
Pillow